requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
python-chess==1.999
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.3.2
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
